Add jack and slide axes (initial)
Move beam path calc axes into new file.
Partially implement bench axes Jack and Slide
