+++ ReflectometryServer/ioc_driver.py
@@ -9,7 +9,7 @@ from pcaspy import Severity
 
 from ReflectometryServer.out_of_beam import OutOfBeamLookup
 from ReflectometryServer.engineering_corrections import NoCorrection, CorrectionUpdate
-from ReflectometryServer.beam_path_calc import DefineValueAsEvent
+from ReflectometryServer.axis import DefineValueAsEvent
 from ReflectometryServer import ChangeAxis
 from ReflectometryServer.pv_wrapper import SetpointUpdate, ReadbackUpdate, IsChangingUpdate
 from ReflectometryServer.server_status_manager import STATUS_MANAGER, ProblemInfo
